driver: sensor: change compass value from int to short
Bug 1266211

Reviewed-on: http://git-master/r/217421
(cherry picked from commit 984398033a9f4c26454e2018ab025f52a115cf91)
Reviewed-on: http://git-master/r/216731
(cherry picked from commit 24cdec613f41532522fc79346e1754bb8bbc183f)
Reviewed-on: http://git-master/r/226355
Tested-by: Xiaohui Tao <xtao@nvidia.com>
GVS: Gerrit_Virtual_Submit

+++ drivers/input/misc/compass/ak8975_input.c
@@ -134,7 +134,7 @@ struct akm_inf {
 	int port_id[2];			/* MPU port ID */
 	u8 data_out;			/* write value to trigger a sample */
 	u8 range_index;			/* max_range index */
-	int xyz[3];			/* sample data */
+	short xyz[3];			/* sample data */
 	bool cycle;			/* cycle MPU en/dis for high speed */
 	unsigned long cycle_delay_us;	/* cycle off time (us) */
 	s64 cycle_ts;			/* cycle MPU disable timestamp */
@@ -548,9 +548,9 @@ static int akm_init_hw(struct akm_inf *inf)
 
 static void akm_calc(struct akm_inf *inf, u8 *data)
 {
-	int x;
-	int y;
-	int z;
+	short x;
+	short y;
+	short z;
 
 	/* data[1] = AKM_REG_HXL
 	 * data[2] = AKM_REG_HXH
@@ -559,9 +559,9 @@ static void akm_calc(struct akm_inf *inf, u8 *data)
 	 * data[5] = AKM_REG_HZL
 	 * data[6] = AKM_REG_HZH
 	 */
-	x = (int)((data[2] << 8) | data[1]);
-	y = (int)((data[4] << 8) | data[3]);
-	z = (int)((data[6] << 8) | data[5]);
+	x = ((data[2] << 8) | data[1]);
+	y = ((data[4] << 8) | data[3]);
+	z = ((data[6] << 8) | data[5]);
 	x = ((x * (inf->asa.asa[AXIS_X] + 128)) >> 8);
 	y = ((y * (inf->asa.asa[AXIS_Y] + 128)) >> 8);
 	z = ((z * (inf->asa.asa[AXIS_Z] + 128)) >> 8);
@@ -981,9 +981,9 @@ static ssize_t akm_magnetic_field_show(struct device *dev,
 				       char *buf)
 {
 	struct akm_inf *inf;
-	int x;
-	int y;
-	int z;
+	short x;
+	short y;
+	short z;
 
 	inf = dev_get_drvdata(dev);
 	if (inf->enable) {
@@ -992,7 +992,7 @@ static ssize_t akm_magnetic_field_show(struct device *dev,
 		y = inf->xyz[AXIS_Y];
 		z = inf->xyz[AXIS_Z];
 		mutex_unlock(&inf->mutex_data);
-		return sprintf(buf, "%d, %d, %d\n", x, y, z);
+		return sprintf(buf, "%hd, %hd, %hd\n", x, y, z);
 	}
 
 	return -EPERM;
